# Film Manager Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  version: 1.0.0
  description: REST APIs that alllows to interact with the Film Manager Service
...
paths:
...
  "/api/films/public":
    get:
      tags: 
      - "api/films/public"
      operationId: getPublicFilms
      summary: retrieve a list of public films.
      description: A generic user can retrieve a list of public films added by other users. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 3,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 2,
                          "title": "Heaven's Feel",
                          "owner": 2,
                          "private": false,
                          "reviews": "/api/films/public/2/reviews",
                          "self": "/api/films/public/2"
                      },
                      {
                          "id": 3,
                          "title": "You Can (Not) Redo",
                          "owner": 1,
                          "private": false,
                          "reviews": "/api/films/public/3/reviews",
                          "self": "/api/films/public/3"
                      }
                  ],
                  "next": "/api/films/public?pageNo=2"
              }
        "500":
          description: Internal Server Error
...
  "/api/films/public/{filmId}/reviews/{reviewerId}/modifications":
    get: 
      tags:
        - api/film/public/{filmId}/reviews/{reviewerId}/modifications
      operationId: getSingleFilmReviewModificationRequests
      security:
        - cookieAuth: []
      summary: retrieve all the review modification request related to a specific review.
      description: An authenticated user can retrieve all the review modification requests
        related to a specific review if he is the owner or the reviewer, ordered for status 
        (pending ones first) and then for mId. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 3
        - name: reviewerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ReviewModifications"
              example: 
                {
                  "totalPages": 2,
                  "currentPage": 1,
                  "mRequests": [
                      {
                          "mId": 1,
                          "filmId": 3,
                          "reviewerId": 1,
                          "deadline": "2026-01-15T14:30:00Z",
                          "status": true,
                          "self": "/api/films/public/reviews/modifications/1",
                          "review": "/api/films/public/3/reviews/1"
                      },
                      {
                          "mId": 2,
                          "filmId": 3,
                          "reviewerId": 1,
                          "deadline": "2026-01-15T14:30:00Z",
                          "status": false,
                          "self": "/api/films/public/reviews/modifications/2",
                          "review": "/api/films/public/3/reviews/1"
                      }
                  ],
                  "next": "/api/films/public/3/reviews/1/modifications?pageNo=2"
                }
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (The user is not owner or the review is not assigned to him)
        "500":
          description: Internal Server Error
    post: 
      tags:
        - api/film/public/{filmId}/reviews/{reviewerId}/modifications
      summary: Used to create a new review modification request
      description: An authenticated user can request to the owner of the film to modify the review.
      security:
        - cookieAuth: []
      operationId: createReviewModificationRequest
      parameters:
      - name: filmId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 6
      - name: reviewerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      requestBody:
        description: Represents the review modification request that is going to be created
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ReviewModification"
            example: 
              {
                "filmId":6,
                "reviewerId":1,
                "deadline":"2026-01-15T14:30:00Z"
            }
        required: true
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ReviewModification"
              example: 
                {
                  "mId": 9,
                  "filmId": 6,
                  "reviewerId": 1,
                  "deadline": "2026-01-15T14:30:00Z",
                  "self": "/api/films/public/reviews/modifications/9",
                  "review": "/api/films/public/6/reviews/1"
                }
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict (Invalid film or reviewerId)
        "500":
          description: Internal Server Error
  "/api/films/public/reviews/modifications/{mId}":
    get: 
      tags:
        - api/film/public/reviews/modifications/{mId}
      operationId: getSingleReviewModificationRequest
      security:
        - cookieAuth: []
      summary: retrieve a specific review modification request.
      description: An authenticated user can retrieve a specific review modification request 
        that he sent to the owner of the public film to modify the review.
      parameters:
        - name: mId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ReviewModification"
              examples :
                example1:
                  description: Accepted review modification request
                  value:
                    {
                      "mId": 1,
                      "filmId": 3,
                      "reviewerId": 1,
                      "deadline": "2026-01-15T14:30:00Z",
                      "status": true,
                      "self": "/api/films/public/reviews/modifications/1",
                      "review": "/api/films/public/3/reviews/1"
                  }
                example2:
                  description: The owner of the film retrieves a pending request of a reviewer.
                  value:
                    {
                      "mId": 7,
                      "filmId": 3,
                      "reviewerId": 2,
                      "deadline": "2026-01-15T14:30:00Z",
                      "status": null,
                      "self": "/api/films/public/reviews/modifications/7",
                      "review": "/api/films/public/3/reviews/2",
                      "accept": "/api/films/public/reviews/modifications/7/accept",
                      "reject": "/api/films/public/reviews/modifications/7/reject"
                    }
                example3:
                  description: The owner of the film retrieves a pending request that he made to himself
                  value:
                    {
                      "mId": 5,
                      "filmId": 3,
                      "reviewerId": 1,
                      "deadline": "2026-01-15T14:30:00Z",
                      "status": null,
                      "self": "/api/films/public/reviews/modifications/5",
                      "review": "/api/films/public/3/reviews/1",
                      "accept": "/api/films/public/reviews/modifications/5/accept",
                      "reject": "/api/films/public/reviews/modifications/5/reject",
                      "delete": "/api/films/public/reviews/modifications/5"
                  }
                example4:
                  description: The reviewer of the film retrieves a pending request.
                  value:
                    {
                        "mId": 9,
                        "filmId": 6,
                        "reviewerId": 1,
                        "deadline": "2026-01-15T14:30:00Z",
                        "status": null,
                        "self": "/api/films/public/reviews/modifications/9",
                        "review": "/api/films/public/6/reviews/1",
                        "delete": "/api/films/public/reviews/modifications/9"
                    }
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    delete:
      tags:
        - api/film/public/reviews/modifications/{mId}
      security:
        - cookieAuth: []
      operationId: deleteSingleReviewModificationRequest
      summary: Delete a specific review modification request.
      description: A logged-in user can delete a specific review modification request 
        that he sent to the film owner while it is in a pending state.
      parameters:
        - name: mId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example : 9
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (The modification request is not in a pending status)
        "500":
          description: Internal Server Error
...
  "/api/films/public/reviews/modifications/status" :
    get: 
      tags:
        - api/films/public/reviews/modifications/status
      operationId: getReviewModificationRequestsStatus
      security:
        - cookieAuth: []
      summary: retrieve the status of all the review modification requests sent.
      description: An authenticated user can retrieve the status 
        of all the review modification request that he sent. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:      
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ReviewModifications"
              example: 
                {
                  "totalPages": 1,
                  "currentPage": 1,
                  "mRequests": [
                      {
                          "mId": 7,
                          "filmId": 3,
                          "reviewerId": 2,
                          "deadline": "2026-01-15T14:30:00Z",
                          "status": null,
                          "self": "/api/films/public/reviews/modifications/7",
                          "review": "/api/films/public/3/reviews/2",
                          "delete": "/api/films/public/reviews/modifications/7"
                      },
                      {
                          "mId": 8,
                          "filmId": 3,
                          "reviewerId": 2,
                          "deadline": "2025-01-16T14:30:00Z",
                          "status": true,
                          "self": "/api/films/public/reviews/modifications/8",
                          "review": "/api/films/public/3/reviews/2"
                      }
                  ],
                  "next": null
              }

        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Film:
      type: object
      description: The schema for the film data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: The path to the schema that describes the film data structure
        id:
          type: integer
          description: The film's unique identifier
        title:
          type: string
          description: The film's title
        owner:
          type: integer
          description: The id of the user that created the film data structure
        private:
          type: boolean
          default: true
          description: It's true if the owner is the only one that can access the film, false otherwise
        watchDate:
          type: string
          format: date
          description: The date when the owner watched the Film
        rating:
          type: integer
          minimum: 0
          maximum: 10
          description: the rating that the owner has given to the film
        favourite:
          type: boolean
          default: false
          description: It's true if the film is among the owner favourites, false otherwise
        reviews:
          type: string
          format: uri
          description: the reviews that other users have given to the film
        self:
          type: string
          format: uri
          description: URI to access the specific film resource.
      required:
        - title
        - owner
        - private
    Films:
      type: object
      additionalProperties: false
      description: the schema for the films data structure
      properties:
        "$schema":
          type: string
          description: The path to the schema used to validate the films data structure
        totalPages:
          type: integer
          minimum: 1
          description: the total number of pages
        currentPage:
          type: integer
          minimum: 1
          description: the number of the currentPage
        films:
          type: array
          items:
            "$ref": "#/components/schemas/Film"
          description: the list of films on the current page
        next:
          type: string
          format: uri
          description: URI for accessing the next page
      required:
        - totalPages
...
    ReviewModification:
      type: object
      description: The schema for the review modification request data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the JSON Schema to use for the validation of the data structure
        mId:
          type: integer
          description: Unique identifier of the review modification request
        filmId:
          type: integer
          description: unique identifier of the film which the review modification request is related 
        reviewerId:
          type: integer
          description: the unique identifier of the user who sent the review modfication request
        deadline:
          type: string
          format: date-time
          description: The deadline by wich an answer is expected
        status:
          type: boolean
          nullable: true
          default: null
          description: Its true if the request is accepted, false if if the request is refused or null if the request is pending
        review:
          type: string
          format: uri
          description: URI to access the review associated with the request.
        accept:
          type: string
          format: uri
          description: URI to accept the modification request.
        reject:
          type: string
          format: uri
          description: URI to reject the modification request.
        delete: 
          type: string
          format: uri
          description: URI to delete the modification request(if in a pending status).
        self:
          type: string
          format: uri
          description: URI to access the specific review modification request.
      required:
        - filmId
        - reviewerId
        - deadline
    ReviewModifications:
      type: object
      description: The schema for the reviewModifications data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: path to the schema used to validate the reviews data structure
        totalPages:
          type: integer
          minimum: 1
          description: the total number of pages
        currentPage:
          type: integer
          minimum: 1
          description: the number of the currentPage
        reviews:
          type: array
          items:
            "$ref": "#/components/schemas/ReviewModification"
          description: the reviews modification requests of the current page
        next:
          type: string
          format: uri
          description: the path to the next page
      required:
        - totalPages
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid